Added unit tests for resource node class.
